improvements: continue tightening of int types, comment hygene, various things surfaced during sweep have been sorted. clang compile warnings dropping significantly

+++ src/_rasterizer_decal_vertices_delete.c
@@ -1,11 +1,11 @@
 #include <stdint.h>
 #include "headers/blam_data_globals.h"
-/* _rasterizer_decal_vertices_delete @0x83711064 — release the cached local-vertex block
+/* _rasterizer_decal_vertices_delete @0x83711058 — release the cached local-vertex block
  * backing a decal's geometry. (Underscore-prefixed implementation; the un-prefixed
  * rasterizer_decal_vertices_delete is the wrapper.) */
 
 #include "headers/lruv_cache.h"
-extern void lruv_block_delete(lruv_cache *cache, uint16_t block_index);
+extern void lruv_block_delete(lruv_cache *cache, int block_index);
 
 void _rasterizer_decal_vertices_delete(int cache_index)
 {
