+++ src/action_charge_update.c
@@ -12,7 +12,7 @@
 #include "headers/blam_data_globals.h"
 
 
-void action_charge_update(uint16_t actor_index)
+void action_charge_update(int actor_index)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
