+++ src/action_converse_perform.c
@@ -20,7 +20,7 @@
 
 
 extern int prop_get_base_by_unit_index(int actor_index, int unit_index, uint8_t acquire, uint8_t refresh_upon_acquisition);
-extern uint8_t actor_move_halt(uint16_t actor_index);
+extern uint8_t actor_move_halt(int actor_index);
 extern uint8_t actor_move_to_prop(int actor_index, unsigned int prop_index, float accept_radius);
 
 uint8_t action_converse_perform(int actor_index)
