+++ src/action_flee_modify_color.c
@@ -6,7 +6,7 @@
 #include "headers/blam_data_globals.h"
 
 
-void action_flee_modify_color(uint16_t actor_index, real_argb_color *color)
+void action_flee_modify_color(int actor_index, real_argb_color *color)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
