+++ src/action_guard_setup_find_position.c
@@ -14,7 +14,7 @@
 #include "headers/blam_data_globals.h"
 
 
-uint8_t action_guard_setup_find_position(uint16_t actor_index, int16_t delay_timer, guard_state_data *state_data)
+uint8_t action_guard_setup_find_position(int actor_index, int16_t delay_timer, guard_state_data *state_data)
 {
     actor_datum *actor = DATA_ARRAY_ELEMENT(actor_data, actor_datum, actor_index);
