+++ src/action_guard_update.c
@@ -21,11 +21,11 @@
 
 
 #include "headers/ai_information_data.h"
-extern int action_guard_speak_post_combat(uint16_t actor_index);
-extern void actor_set_dormant(uint16_t actor_index, uint8_t dormant);
-extern int actor_target_unit_index(uint16_t actor_index);
+extern int action_guard_speak_post_combat(int actor_index);
+extern void actor_set_dormant(int actor_index, uint8_t dormant);
+extern int actor_target_unit_index(int actor_index);
 extern void ai_communication_event(int16_t communication_type, int subject_unit_index, int cause_unit_index, int16_t hostility, int16_t damage_type, int16_t information_type, ai_information_data *information_data);
-extern void actor_discard_firing_position(uint16_t actor_index, int16_t firing_position_index, uint8_t temporary);
+extern void actor_discard_firing_position(int actor_index, int16_t firing_position_index, uint8_t temporary);
 
 void action_guard_update(int actor_index)
 {
