+++ src/action_obey_command_list_setup.c
@@ -19,7 +19,7 @@
 #include "headers/blam_data_globals.h"
 #include "headers/obey_individual_iterator.h"
 
-extern void actor_look_secondary_stop(uint16_t actor_index);
+extern void actor_look_secondary_stop(int actor_index);
 
 extern void action_obey_individual_setup(int actor_index, int unit_index, int16_t command_list_index, obey_individual_simple_control *simple_control, obey_individual_complex_control *complex_control, uint8_t *user_data);
 extern void action_obey_individuals_iterate(int actor_index, uint8_t initialize_structures, obey_state_data *state_data, obey_individual_iterator iterator, void *user_data);
